requirements.txt
flask==3.0.3
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.1
matplotlib==3.9.0
joblib==1.4.2
pyyaml==6.0.2
requests==2.32.3
pytest==8.3.2
Werkzeug==3.0.3
